ACS-1110 replace fabric8 with docker-maven-plugin (#1252)
The community repo image was not being updated on DockerHub or quay.io after the original commit for this issue. It appears the new io.fabric8:docker-maven-plugin is only able to push to one registry. We had images for quay.io and dockerhub.

On a branch we managed to release 7.0.0-A27c. In order to do so we temporarily had to comment out the upstream dependency check in build.sh.

[skip ci]